generate data/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.2=py36hecd8cb5_1001
argon2-cffi=20.1.0=py36h9ed2024_1
async_generator=1.10=py36h28b3542_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
brotli=1.0.9=hb1e8313_2
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.4.13=hecd8cb5_1
cairo=1.16.0=h8023c5d_1
certifi=2020.12.5=py36hecd8cb5_0
cffi=1.14.5=py36h2125817_0
decorator=5.0.6=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py36_0
ffmpeg=4.0=h01ea3c9_0
fontconfig=2.13.1=ha9ee91d_0
fonttools=4.22.1=pyhd3eb1b0_0
freetype=2.10.4=ha233b18_0
gettext=0.21.0=h7535e17_0
glib=2.68.1=hdf23fa2_0
graphite2=1.3.14=h38d11af_0
harfbuzz=1.8.8=hb8d4a28_0
hdf5=1.10.2=hfa1e0ec_1
icu=58.2=h0a44026_3
importlib-metadata=3.10.0=py36hecd8cb5_0
importlib_metadata=3.10.0=hd3eb1b0_0
intel-openmp=2021.2.0=hecd8cb5_564
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jasper=2.0.14=h636a363_1
jedi=0.17.0=py36_0
jinja2=2.11.3=pyhd3eb1b0_0
jpeg=9b=he5867d9_2
jsonschema=3.2.0=py_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py36hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
lcms2=2.12=hf1fd2bf_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libopencv=3.4.2=h7c891bd_1
libopus=1.3.1=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.1.0=hcb84e12_1
libvpx=1.7.0=h378b8a2_0
libxml2=2.9.10=h7cdb67c_3
llvm-openmp=10.0.0=h28b9765_0
lz4-c=1.9.3=h23ab428_0
markupsafe=1.1.1=py36h1de35cc_0
mistune=0.8.4=py36h1de35cc_0
mkl=2019.4=233
mkl-service=2.3.0=py36h9ed2024_0
mkl_fft=1.3.0=py36ha059aab_0
mkl_random=1.1.1=py36h959d312_0
munkres=1.1.4=py_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py36_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.3.0=py36hecd8cb5_0
numpy=1.19.2=py36h456fd55_0
numpy-base=1.19.2=py36hcfb5961_0
olefile=0.46=py36_0
opencv=3.4.2=py36h6fd60c2_1
openssl=1.1.1k=h9ed2024_0
packaging=20.9=pyhd3eb1b0_0
pandoc=2.12=hecd8cb5_0
pandocfilters=1.4.3=py36hecd8cb5_1
parso=0.8.2=pyhd3eb1b0_0
pcre=8.44=hb1e8313_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py36h5270095_0
pip=21.0.1=py36hecd8cb5_0
pixman=0.40.0=haf1e3a3_0
prometheus_client=0.10.1=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py-opencv=3.4.2=py36h7c891bd_1
pycparser=2.20=py_2
pygments=2.8.1=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyrsistent=0.17.3=py36haf1e3a3_0
python=3.6.13=h88f2d9e_0
python-dateutil=2.8.1=pyhd3eb1b0_0
pyzmq=20.0.0=py36h23ab428_1
readline=8.1=h9ed2024_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py36hecd8cb5_0
six=1.15.0=py36hecd8cb5_0
sqlite=3.35.4=hce871da_0
terminado=0.9.4=py36hecd8cb5_0
testpath=0.4.4=pyhd3eb1b0_0
tk=8.6.10=hb0a8c7a_0
tornado=6.1=py36h9ed2024_0
traitlets=4.3.3=py36_0
typing_extensions=3.7.4.3=pyha847dfd_0
unicodedata2=13.0.0.post2=py36h9ed2024_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py36_1
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zeromq=4.3.4=h23ab428_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
